add schematic
